product/juno: dts: add devicetree support for clock and mock clock modules
Add support to juno build and source for clock configured from devicetree.
