+++ product/juno/scp_ramfw/CMakeLists.txt
@@ -17,15 +17,19 @@ target_include_directories(
 
 # cmake-lint: disable=E1122
 
+# files removed from build and replaced with _dt_ build versions
+# as modules added to dt vars in Firmware.cmake
 #            "${CMAKE_CURRENT_SOURCE_DIR}/config_pl011.c"
+#             "${CMAKE_CURRENT_SOURCE_DIR}/config_clock.c"
+#             "${CMAKE_CURRENT_SOURCE_DIR}/config_mock_clock.c"
 
 
 target_sources(
     juno-bl2
     PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/config_sds.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/config_dt_juno_pl011.c"
+            "${CMAKE_CURRENT_SOURCE_DIR}/config_dt_juno_clock.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/config_juno_soc_clock_ram.c"
-            "${CMAKE_CURRENT_SOURCE_DIR}/config_clock.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/config_dvfs.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/config_juno_cdcel937.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/config_juno_hdlcd.c"
@@ -52,7 +56,6 @@ target_sources(
             "${CMAKE_CURRENT_SOURCE_DIR}/config_juno_pvt.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/config_juno_thermal.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/config_scmi_power_domain.c"
-            "${CMAKE_CURRENT_SOURCE_DIR}/config_mock_clock.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/../src/juno_utils.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/../src/juno_id.c"
             "${CMAKE_CURRENT_SOURCE_DIR}/../src/juno_scmi_clock.c")
