+++ product/juno/scp_ramfw/deprecated/config_mock_clock.c
